+++ lib/git_cli/test/test__num_status.ml
@@ -19,16 +19,9 @@
 (*  <http://www.gnu.org/licenses/> and <https://spdx.org>, respectively.       *)
 (*******************************************************************************)
 
-(* [super-master-mind.num-status] has been created by capturing the output of:
+(* In this test we verify that we can parse chosen outputs from the data directory. *)
 
-   {v
-      $ git diff --numstat 1892d4980ee74945eb98f67be26b745f96c0f482..bcaf94757fe3cb247fa544445f0f41f3616943d7
-   v}
-
-   In this test we verify that we can parse this output.contents
-*)
-
-let%expect_test "parse_exn" =
+let%expect_test "parse_exn - super-master-mind" =
   Eio_main.run
   @@ fun env ->
   let path = Eio.Path.(Eio.Stdenv.fs env / "super-master-mind.num-status") in
@@ -39,239 +32,425 @@ let%expect_test "parse_exn" =
   [%expect
     {|
     (((key (One_file .github/workflows/ci.yml))
-      (num_lines_in_diff (
-        (insertions 3)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 3)
+          (deletions  2)))))
      ((key (One_file .github/workflows/deploy-odoc.yml))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  28))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  28)))))
      ((key (One_file .vscode/settings.json))
-      (num_lines_in_diff (
-        (insertions 2)
-        (deletions  0))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 2)
+          (deletions  0)))))
      ((key (One_file CHANGELOG.md))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  28))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  28)))))
      ((key (One_file CHANGES.md))
-      (num_lines_in_diff (
-        (insertions 54)
-        (deletions  0))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 54)
+          (deletions  0)))))
      ((key (One_file CODE_OF_CONDUCT.md))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  128))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  128)))))
      ((key (One_file Makefile))
-      (num_lines_in_diff (
-        (insertions 4)
-        (deletions  1))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 4)
+          (deletions  1)))))
      ((key (One_file README.md))
-      (num_lines_in_diff (
-        (insertions 63)
-        (deletions  59))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 63)
+          (deletions  59)))))
      ((key (One_file bin/dune))
-      (num_lines_in_diff (
-        (insertions 2)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 2)
+          (deletions  2)))))
      ((key (One_file bin/main.ml))
-      (num_lines_in_diff (
-        (insertions 1)
-        (deletions  1))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 1)
+          (deletions  1)))))
      ((key (One_file dune-project))
-      (num_lines_in_diff (
-        (insertions 72)
-        (deletions  3))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 72)
+          (deletions  3)))))
      ((key (One_file lib/super_master_mind/src/code.ml))
-      (num_lines_in_diff (
-        (insertions 11)
-        (deletions  7))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 11)
+          (deletions  7)))))
      ((key (One_file lib/super_master_mind/src/code.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/codes.ml))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  3))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  3)))))
      ((key (One_file lib/super_master_mind/src/codes.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/color.ml))
-      (num_lines_in_diff (
-        (insertions 15)
-        (deletions  6))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 15)
+          (deletions  6)))))
      ((key (One_file lib/super_master_mind/src/color.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/color_permutation.ml))
-      (num_lines_in_diff (
-        (insertions 4)
-        (deletions  5))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 4)
+          (deletions  5)))))
      ((key (One_file lib/super_master_mind/src/color_permutation.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/cue.ml))
-      (num_lines_in_diff (
-        (insertions 4)
-        (deletions  6))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 4)
+          (deletions  6)))))
      ((key (One_file lib/super_master_mind/src/cue.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/dune))
-      (num_lines_in_diff (
-        (insertions 18)
-        (deletions  6))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 18)
+          (deletions  6)))))
      ((key (One_file lib/super_master_mind/src/example.ml))
-      (num_lines_in_diff (
-        (insertions 1)
-        (deletions  5))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 1)
+          (deletions  5)))))
      ((key (One_file lib/super_master_mind/src/example.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/game_dimensions.ml))
-      (num_lines_in_diff (
-        (insertions 11)
-        (deletions  5))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 11)
+          (deletions  5)))))
      ((key (One_file lib/super_master_mind/src/game_dimensions.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  3))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  3)))))
      ((key (One_file lib/super_master_mind/src/guess.ml))
-      (num_lines_in_diff (
-        (insertions 73)
-        (deletions  44))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 73)
+          (deletions  44)))))
      ((key (One_file lib/super_master_mind/src/guess.mli))
-      (num_lines_in_diff (
-        (insertions 25)
-        (deletions  5))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 25)
+          (deletions  5)))))
      ((key (One_file lib/super_master_mind/src/import/dune))
-      (num_lines_in_diff (
-        (insertions 9)
-        (deletions  0))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 9)
+          (deletions  0)))))
      ((key (
         Two_files
         (src lib/super_master_mind/src/import.ml)
         (dst lib/super_master_mind/src/import/super_master_mind_import.ml)))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (
         Two_files
         (src lib/super_master_mind/src/import.mli)
         (dst lib/super_master_mind/src/import/super_master_mind_import.mli)))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/kheap.ml))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  3))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  3)))))
      ((key (One_file lib/super_master_mind/src/kheap.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/maker.ml))
-      (num_lines_in_diff (
-        (insertions 3)
-        (deletions  6))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 3)
+          (deletions  6)))))
      ((key (One_file lib/super_master_mind/src/maker.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/opening_book.ml))
-      (num_lines_in_diff (
-        (insertions 48)
-        (deletions  22))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 48)
+          (deletions  22)))))
      ((key (One_file lib/super_master_mind/src/opening_book.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/solver.ml))
-      (num_lines_in_diff (
-        (insertions 1)
-        (deletions  4))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 1)
+          (deletions  4)))))
      ((key (One_file lib/super_master_mind/src/solver.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/super_master_mind.ml))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/super_master_mind.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  1))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  1)))))
      ((key (One_file lib/super_master_mind/src/task_pool.ml))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/src/task_pool.mli))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  2))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
      ((key (One_file lib/super_master_mind/test/dune))
-      (num_lines_in_diff (
-        (insertions 18)
-        (deletions  3))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 18)
+          (deletions  3)))))
      ((key (One_file lib/super_master_mind/test/test__code.ml))
-      (num_lines_in_diff (
-        (insertions 85)
-        (deletions  33))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 85)
+          (deletions  33)))))
      ((key (One_file lib/super_master_mind/test/test__codes.ml))
-      (num_lines_in_diff (
-        (insertions 92)
-        (deletions  31))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 92)
+          (deletions  31)))))
      ((key (One_file lib/super_master_mind/test/test__color.ml))
-      (num_lines_in_diff (
-        (insertions 7)
-        (deletions  7))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 7)
+          (deletions  7)))))
      ((key (One_file lib/super_master_mind/test/test__color_permutation.ml))
-      (num_lines_in_diff (
-        (insertions 5)
-        (deletions  11))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 5)
+          (deletions  11)))))
      ((key (One_file lib/super_master_mind/test/test__cue.ml))
-      (num_lines_in_diff (
-        (insertions 50)
-        (deletions  34))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 50)
+          (deletions  34)))))
      ((key (One_file lib/super_master_mind/test/test__example.ml))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  3))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  3)))))
      ((key (One_file lib/super_master_mind/test/test__game_dimensions.ml))
-      (num_lines_in_diff (
-        (insertions 9)
-        (deletions  11))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 9)
+          (deletions  11)))))
      ((key (One_file lib/super_master_mind/test/test__guess.ml))
-      (num_lines_in_diff (
-        (insertions 678)
-        (deletions  278))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 678)
+          (deletions  278)))))
      ((key (One_file lib/super_master_mind/test/test__kheap.ml))
-      (num_lines_in_diff (
-        (insertions 0)
-        (deletions  4))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  4)))))
      ((key (One_file lib/super_master_mind/test/test__mins.ml))
-      (num_lines_in_diff (
-        (insertions 108)
-        (deletions  64))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 108)
+          (deletions  64)))))
      ((key (One_file lib/super_master_mind/test/test__opening_book.ml))
-      (num_lines_in_diff (
-        (insertions 9)
-        (deletions  11))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 9)
+          (deletions  11)))))
      ((key (One_file super-master-mind.opam))
-      (num_lines_in_diff (
-        (insertions 32)
-        (deletions  22))))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 32)
+          (deletions  22)))))
      ((key (One_file test/maker.t))
-      (num_lines_in_diff (
-        (insertions 4)
-        (deletions  5))))) |}];
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 4)
+          (deletions  5)))))) |}];
+  ()
+;;
+
+let%expect_test "parse_exn - eio" =
+  Eio_main.run
+  @@ fun env ->
+  let path = Eio.Path.(Eio.Stdenv.fs env / "eio.num-status") in
+  let contents = Eio.Path.load path in
+  let lines = String.split_lines contents in
+  let num_status = Git_cli.Num_status.parse_lines_exn ~lines in
+  print_s [%sexp (num_status : Vcs.Num_status.t)];
+  [%expect
+    {|
+    (((key (One_file CHANGES.md))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 31)
+          (deletions  0)))))
+     ((key (One_file README.md))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 106)
+          (deletions  20)))))
+     ((key (One_file bench/main.ml))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 13)
+          (deletions  0)))))
+     ((key (One_file doc/multicore.md))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 32)
+          (deletions  9)))))
+     ((key (One_file doc/traces/Makefile))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 1)
+          (deletions  1)))))
+     ((key (One_file doc/traces/multicore-posix.fxt)) (num_stat Binary_file))
+     ((key (One_file doc/traces/multicore-posix.svg))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 1490)
+          (deletions  0)))))
+     ((key (One_file dune-project))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 5)
+          (deletions  5)))))
+     ((key (One_file eio.opam))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 1)
+          (deletions  1)))))
+     ((key (One_file eio_linux.opam))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 2)
+          (deletions  2)))))
+     ((key (One_file eio_main.opam))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 1)
+          (deletions  1)))))
+     ((key (One_file eio_posix.opam))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 1)
+          (deletions  1)))))
+     ((key (One_file lib_eio/pool.ml))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 29)
+          (deletions  9)))))
+     ((key (One_file lib_eio/pool.mli))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 8)
+          (deletions  2)))))
+     ((key (One_file lib_eio_linux/dune))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 1)
+          (deletions  1)))))
+     ((key (One_file lib_eio_linux/eio_linux.ml))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 10)
+          (deletions  10)))))
+     ((key (One_file lib_eio_linux/eio_linux.mli))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 1)
+          (deletions  204)))))
+     ((key (One_file lib_eio_linux/log.ml))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 0)
+          (deletions  2)))))
+     ((key (One_file lib_eio_linux/low_level.ml))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 8)
+          (deletions  13)))))
+     ((key (One_file lib_eio_linux/low_level.mli))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 242)
+          (deletions  0)))))
+     ((key (One_file lib_eio_linux/sched.ml))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 7)
+          (deletions  11)))))
+     ((key (One_file lib_eio_linux/tests/dune))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 2)
+          (deletions  2)))))
+     ((key (One_file lib_eio_linux/tests/test.ml))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 5)
+          (deletions  5)))))
+     ((key (One_file tests/pool.md))
+      (num_stat (
+        Num_lines_in_diff (
+          (insertions 31)
+          (deletions  0)))))) |}];
   ()
 ;;
 
@@ -285,6 +464,7 @@ let%expect_test "parse_lines_exn" =
     ; "0\t1\tfile"
     ; "1985\t0\tfile1 => file2"
     ; "100\t5\ttmp/{dir1 => dir2}/file"
+    ; "-\t-\tfile"
     ]
   in
   List.iter lines ~f:(fun line ->
@@ -300,26 +480,30 @@ let%expect_test "parse_lines_exn" =
     ("0\t1\tfile" (
       Ok (
         (key (One_file file))
-        (num_lines_in_diff (
-          (insertions 0)
-          (deletions  1))))))
+        (num_stat (
+          Num_lines_in_diff (
+            (insertions 0)
+            (deletions  1)))))))
     ("1985\t0\tfile1 => file2" (
       Ok (
         (key (
           Two_files
           (src file1)
           (dst file2)))
-        (num_lines_in_diff (
-          (insertions 1985)
-          (deletions  0))))))
+        (num_stat (
+          Num_lines_in_diff (
+            (insertions 1985)
+            (deletions  0)))))))
     ("100\t5\ttmp/{dir1 => dir2}/file" (
       Ok (
         (key (
           Two_files
           (src tmp/dir1/file)
           (dst tmp/dir2/file)))
-        (num_lines_in_diff (
-          (insertions 100)
-          (deletions  5)))))) |}];
+        (num_stat (
+          Num_lines_in_diff (
+            (insertions 100)
+            (deletions  5)))))))
+    ("-\t-\tfile" (Ok ((key (One_file file)) (num_stat Binary_file)))) |}];
   ()
 ;;